split test suite by endpoint domain
- Split monolithic spotify_test.dart into 16 separate files
- Each file corresponds to a specific endpoint/domain
- All 44 tests passing
- Improved test organization and maintainability
