+++ test/errors_test.dart
@@ -1,0 +1,50 @@
+import 'dart:async';
+import 'spotify_mock.dart';
+import 'package:test/test.dart';
+import 'package:spotify/spotify.dart';
+
+// ignore_for_file: deprecated_member_use_from_same_package
+
+Future main() async {
+  var spotify = SpotifyApiMock(SpotifyApiCredentials(
+    'clientId',
+    'clientSecret',
+  ));
+
+  tearDown(() {
+    spotify.interceptor = null;
+    spotify.mockHttpErrors = <MockHttpError>[].iterator;
+  });
+
+  group('Errors', () {
+    test('apiRateErrorSuccess', () async {
+      spotify.mockHttpErrors = List.generate(
+          4,
+          (i) => MockHttpError(
+              statusCode: 429,
+              message: 'API Rate exceeded',
+              headers: {'retry-after': '1'})).iterator;
+      var artist = await spotify.artists.get('0TnOYISbd1XYRBk9myaseg');
+      expect(artist.type, 'artist');
+      expect(artist.id, '0TnOYISbd1XYRBk9myaseg');
+      expect(artist.images!.length, 3);
+    });
+    test('apiRateErrorFail', () async {
+      spotify.mockHttpErrors = List.generate(
+          10,
+          (i) => MockHttpError(
+              statusCode: 429,
+              message: 'API Rate exceeded',
+              headers: {'retry-after': '1'})).iterator;
+      late ApiRateException ex;
+      try {
+        await spotify.artists.get('0TnOYISbd1XYRBk9myaseg');
+      } on ApiRateException catch (e) {
+        expect(e, isA<ApiRateException>());
+        ex = e;
+      }
+      expect(ex, isNotNull);
+      expect(ex.status, 429);
+    });
+  });
+}
