+++ test/search_test.dart
@@ -1,0 +1,40 @@
+import 'dart:async';
+import 'spotify_mock.dart';
+import 'package:test/test.dart';
+import 'package:spotify/spotify.dart';
+
+// ignore_for_file: deprecated_member_use_from_same_package
+
+Future main() async {
+  var spotify = SpotifyApiMock(SpotifyApiCredentials(
+    'clientId',
+    'clientSecret',
+  ));
+
+  tearDown(() {
+    spotify.interceptor = null;
+    spotify.mockHttpErrors = <MockHttpError>[].iterator;
+  });
+
+  group('Search', () {
+    test('get', () async {
+      var searchResult = await spotify.search.get('metallica').first();
+      expect(searchResult.length, 2);
+    });
+
+    test('getError', () async {
+      spotify.mockHttpErrors =
+          [MockHttpError(statusCode: 401, message: 'Bad Request')].iterator;
+      late SpotifyException ex;
+      try {
+        await spotify.search.get('metallica').first();
+      } on SpotifyException catch (e) {
+        expect(e, isA<SpotifyException>());
+        ex = e;
+      }
+      expect(ex, isNotNull);
+      expect(ex.status, 401);
+      expect(ex.message, 'Bad Request');
+    });
+  });
+}
